requirements.txt
torch
matplotlib
tensorboard
networkx
pyyaml
numpy
